1. CProfile::nVersion : int -> uint16 2. CMvPeer::nVersion : int -> uint32 3. MV_VERSION, PROTO_VERSION : int -> uint32

+++ common/profile.h
@@ -26,7 +26,7 @@ enum
 class CProfile
 {
 public:
-    int nVersion;
+    uint16 nVersion;
     std::string strName;
     std::string strSymbol;
     uint8 nFlag;
